requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.11.5
pydantic-settings==2.9.1
pyjwt==2.8.0
paho-mqtt==1.6.1
twilio==9.1.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
